backend/requirements.txt
Django>=5.0
gunicorn
djangorestframework>=3.14.0
psycopg2-binary
python-dotenv>=1.0.0
dj-database-url
django-cors-headers>=4.3.0
